Various fixes for the previous commit:
Changed Integer to BigInteger for the Town.fips field so postgres instance on the staging server can save longer intgers in that field.
Removed some filters from the indicator creation dialog.

+++ ckanext/ctdata_theme/ctdata/community/models.py
@@ -24,7 +24,7 @@ class CommunityProfile(Base):
 class Town(Base):
     __tablename__ = 'ctdata_towns'
 
-    fips = Column(Integer, primary_key=True)
+    fips = Column(BigInteger, primary_key=True)
     name = Column(String)
 
     def __init__(self, fips, name):
